+++ package.json
@@ -9,6 +9,7 @@
     "lucide-react": "^0.263.1",
     "react": "^18.2.0",
     "react-dom": "^18.2.0",
+    "react-feather": "^2.0.10",
     "react-scripts": "5.0.1",
     "web-vitals": "^2.1.4"
   },
